Rename RouteEntry to AgentRoute Split the route_common.cc to files corresponding to each class.
