+++ src/vnsw/agent/cfg/cfg_init.cc
@@ -38,7 +38,7 @@
 #include <oper/nexthop.h>
 #include <oper/interface_common.h>
 #include <oper/mirror_table.h>
-#include <oper/agent_route.h>
+#include <oper/route_common.h>
 #include <oper/operdb_init.h>
 #include <oper/global_vrouter.h>
