+++ src/vnsw/agent/oper/SConscript
@@ -26,10 +26,12 @@ for src in SandeshGenSrcs:
 vnswoperdb = env.Library('vnswoperdb',
                          sandesh_objs +
                          ['agent_sandesh.cc',
+                          'agent_path.cc',
+                          'agent_route.cc',
                           'global_vrouter.cc',
                           'inet_interface.cc',
-                          'inet4_mcroute.cc',
-                          'inet4_ucroute.cc',
+                          'inet4_multicast_route.cc',
+                          'inet4_unicast_route.cc',
                           'interface.cc',
                           'layer2_route.cc',
                           'mirror_table.cc',
@@ -38,7 +40,6 @@ vnswoperdb = env.Library('vnswoperdb',
                           'nexthop.cc',
                           'operdb_init.cc',
                           'peer.cc',
-                          'route_common.cc', 
                           'sg.cc',
                           'vm.cc',
                           'vm_interface.cc',
