+++ src/vnsw/agent/oper/agent_route.h
@@ -29,7 +29,7 @@
 
 //Route entry in route table related classes
 
-class RouteEntry : public Route {
+class AgentRoute : public Route {
 public:
     enum Trace {
         ADD,
@@ -39,16 +39,16 @@ public:
         CHANGE_PATH,
     };
 
-    RouteEntry(VrfEntry *vrf, bool is_multicast) : Route(), vrf_(vrf), 
+    AgentRoute(VrfEntry *vrf, bool is_multicast) : Route(), vrf_(vrf), 
         is_multicast_(is_multicast) { };
-    virtual ~RouteEntry() { };
+    virtual ~AgentRoute() { };
 
     //TODO Rename iterator as gw_route-ITERATOR
-    typedef DependencyList<RouteEntry, RouteEntry>::iterator iterator;
-    typedef DependencyList<RouteEntry, RouteEntry>::const_iterator
+    typedef DependencyList<AgentRoute, AgentRoute>::iterator iterator;
+    typedef DependencyList<AgentRoute, AgentRoute>::const_iterator
         const_iterator;
-    typedef DependencyList<NextHop, RouteEntry>::iterator tunnel_nh_iterator;
-    typedef DependencyList<NextHop, RouteEntry>::const_iterator
+    typedef DependencyList<NextHop, AgentRoute>::iterator tunnel_nh_iterator;
+    typedef DependencyList<NextHop, AgentRoute>::const_iterator
         const_tunnel_nh_iterator;
 
     virtual int CompareTo(const Route &rhs) const = 0;
@@ -98,42 +98,39 @@ private:
 
     VrfEntryRef vrf_;
     bool is_multicast_;
-    DEPENDENCY_LIST(RouteEntry, RouteEntry, dependant_routes_);
-    DEPENDENCY_LIST(NextHop, RouteEntry, tunnel_nh_list_);
-    DISALLOW_COPY_AND_ASSIGN(RouteEntry);
+    DEPENDENCY_LIST(AgentRoute, AgentRoute, dependant_routes_);
+    DEPENDENCY_LIST(NextHop, AgentRoute, tunnel_nh_list_);
+    DISALLOW_COPY_AND_ASSIGN(AgentRoute);
 };
 
 class RouteComparator {
 public:
-    bool operator() (const RouteEntry *rt1, const RouteEntry *rt2) {
+    bool operator() (const AgentRoute *rt1, const AgentRoute *rt2) {
         return rt1->IsLess(*rt2);
     }
 };
 
-class NHComparator {
-public:
+struct NHComparator {
     bool operator() (const NextHop *nh1, const NextHop *nh2) {
         return nh1->IsLess(*nh2);
     }
 };
 
-class RouteTableWalkerState {
-public:
+struct RouteTableWalkerState {
     RouteTableWalkerState(LifetimeActor *actor): rt_delete_ref_(this, actor) {
     }
     ~RouteTableWalkerState() {
         rt_delete_ref_.Reset(NULL);
     }
     void ManagedDelete() { };
-private:
     LifetimeRef<RouteTableWalkerState> rt_delete_ref_;
 };
 
 class AgentRouteTable : public RouteTable {
 public:
-    typedef set<const RouteEntry *, RouteComparator> UnresolvedRouteTree;
+    typedef set<const AgentRoute *, RouteComparator> UnresolvedRouteTree;
     typedef set<const NextHop *, NHComparator> UnresolvedNHTree;
-    typedef set<const RouteEntry *, RouteComparator>::const_iterator 
+    typedef set<const AgentRoute *, RouteComparator>::const_iterator 
         const_rt_iterator;
     typedef set<const NextHop *, NHComparator>::const_iterator const_nh_iterator;
 
@@ -145,8 +142,8 @@ public:
     virtual size_t Hash(const DBEntry *entry) const {return 0;};
     virtual size_t Hash(const DBRequestKey *key) const {return 0;};
 
-    virtual void ProcessDelete(RouteEntry *rt) { };
-    virtual void ProcessAdd(RouteEntry *rt) { };
+    virtual void ProcessDelete(AgentRoute *rt) { };
+    virtual void ProcessAdd(AgentRoute *rt) { };
     virtual void RouteTableWalkerNotify(VrfEntry *vrf, AgentXmppChannel *, 
                                         DBState *, bool associate,
                                         bool unicast_walk, bool multicast_walk);
@@ -164,13 +161,13 @@ public:
     void UnicastRouteNotifyDone(DBTableBase *base, DBState *, Peer *);
     void MulticastRouteNotifyDone(DBTableBase *base, DBState *, Peer *);
 
-    void AddUnresolvedRoute(const RouteEntry *rt);
-    void RemoveUnresolvedRoute(const RouteEntry *rt);
+    void AddUnresolvedRoute(const AgentRoute *rt);
+    void RemoveUnresolvedRoute(const AgentRoute *rt);
     void EvaluateUnresolvedRoutes(void);
     void AddUnresolvedNH(const NextHop *);
     void RemoveUnresolvedNH(const NextHop *);
     void EvaluateUnresolvedNH(void);
-    RouteEntry *FindActiveEntry(const RouteKey *key);
+    AgentRoute *FindActiveEntry(const RouteKey *key);
     NextHop *FindNextHop(NextHopKey *key) const;
     bool DelExplicitRoute(DBTablePartBase *part, DBEntryBase *entry);
 
@@ -209,7 +206,7 @@ public:
 private:
     class DeleteActor;
     void Input(DBTablePartition *part, DBClient *client, DBRequest *req);
-    void DeleteRoute(DBTablePartBase *part, RouteEntry *rt, 
+    void DeleteRoute(DBTablePartBase *part, AgentRoute *rt, 
                      const Peer *peer);
     UnresolvedRouteTree unresolved_rt_tree_;
     UnresolvedNHTree unresolved_nh_tree_;
@@ -221,90 +218,20 @@ private:
     DISALLOW_COPY_AND_ASSIGN(AgentRouteTable);
 };
 
-// Path info for every route entry
-
-class AgentPath : public Path {
-public:
-    AgentPath(const Peer *peer, RouteEntry *rt) : 
-        Path(), peer_(peer), nh_(NULL), label_(MplsTable::kInvalidLabel),
-        dest_vn_name_(""), unresolved_(true), sync_(false), vrf_name_(""), 
-        dependant_rt_(rt), proxy_arp_(false), force_policy_(false),
-        tunnel_bmap_(0), interfacenh_flags_(0) { };
-    virtual ~AgentPath() { }; 
-
-    const Peer *GetPeer() const {return peer_;};
-    const NextHop *GetNextHop() const; 
-    uint32_t GetLabel() const {return label_;};
-    int GetTunnelBmap() const {return tunnel_bmap_;};
-    TunnelType::Type GetTunnelType() const {
-        return TunnelType::ComputeType(tunnel_bmap_);
-    };
-    const string &GetDestVnName() const {return dest_vn_name_;};
-    const Ip4Address& GetGatewayIp() const {return gw_ip_;};
-    const string &GetVrfName() const {return vrf_name_;};
-    bool GetProxyArp() const {return proxy_arp_;};
-    bool GetForcePolicy() const {return force_policy_;};
-    const bool IsUnresolved() const {return unresolved_;};
-    uint8_t GetInterfaceNHFlags() const {return interfacenh_flags_;};
-    const SecurityGroupList &GetSecurityGroupList() const {return sg_list_;};  
-
-    void SetLabel(uint32_t label) {label_ = label;};
-    void SetDestVnName(const string &dest_vn) {dest_vn_name_ = dest_vn;};
-    void SetUnresolved(bool unresolved) {unresolved_ = unresolved;};
-    void SetGatewayIp(const Ip4Address &addr) {gw_ip_ = addr;};
-    void SetProxyArp(bool proxy_arp) {proxy_arp_ = proxy_arp;};
-    void SetForcePolicy(bool force_policy) {force_policy_ = force_policy;};
-    void SetVrfName(const string &vrf_name) {vrf_name_ = vrf_name;};
-    void SetTunnelBmap(TunnelType::TypeBmap bmap) {tunnel_bmap_ = bmap;};
-    void SetInterfaceNHFlags(uint8_t flags) {interfacenh_flags_ = flags;};
-    void SetSecurityGroupList(SecurityGroupList &sg) {sg_list_ = sg;};
-
-    void ClearSecurityGroupList() { sg_list_.clear(); }
-    void ResetDependantRoute(RouteEntry *rt) {dependant_rt_.reset(rt);};
-
-    bool ChangeNH(NextHop *nh);
-    bool Sync(RouteEntry *sync_route); //vm_path sync
-    void SyncRoute(bool sync) {sync_ = sync;};
-    bool RouteNeedsSync() {return sync_;};
-    virtual std::string ToString() const { return "AgentPath"; };
-
-private:
-    const Peer *peer_;
-    NextHopRef nh_;
-    uint32_t label_;
-    string dest_vn_name_;
-    // Points to gateway route, if this path is part of
-    // indirect route
-    bool unresolved_;
-    bool sync_;
-    string vrf_name_;
-    Ip4Address gw_ip_;
-    DependencyRef<RouteEntry, RouteEntry> dependant_rt_;
-    bool proxy_arp_;
-    bool force_policy_;
-    TunnelType::TypeBmap tunnel_bmap_;
-    uint8_t interfacenh_flags_;
-    SecurityGroupList sg_list_;
-    DISALLOW_COPY_AND_ASSIGN(AgentPath);
-};
-
-// Route Key related classes 
-
-class RouteKey : public AgentKey {
-public:
+struct RouteKey : public AgentKey {
     RouteKey(const Peer *peer, const string &vrf_name) : AgentKey(), 
     peer_(peer), vrf_name_(vrf_name) { };
     virtual ~RouteKey() { };
 
     virtual AgentRouteTable *GetRouteTableFromVrf(VrfEntry *vrf) = 0; 
     virtual AgentRouteTableAPIS::TableType GetRouteTableType() = 0;
-    virtual RouteEntry *AllocRouteEntry(VrfEntry *vrf, 
+    virtual AgentRoute *AllocRouteEntry(VrfEntry *vrf, 
                                         bool is_multicast) const = 0;
     virtual string ToString() const = 0;
 
     const string &GetVrfName() const { return vrf_name_; };
     const Peer *GetPeer() const { return peer_; };
-private:
+
     const Peer *peer_;
     string vrf_name_;
     DISALLOW_COPY_AND_ASSIGN(RouteKey);
@@ -312,7 +239,7 @@ private:
 
 //Route data related classes
 
-class RouteData : public AgentData {
+struct RouteData : public AgentData {
 public:
     enum Op {
         CHANGE,
@@ -331,786 +258,6 @@ private:
     Op op_;
     bool is_multicast_;
     DISALLOW_COPY_AND_ASSIGN(RouteData);
-};
-
-class ResolveRoute : public RouteData {
-public:
-    ResolveRoute(Op op  = RouteData::CHANGE) : RouteData(op, false) { };
-    virtual ~ResolveRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "Resolve";};;
-private:
-    DISALLOW_COPY_AND_ASSIGN(ResolveRoute);
-};
-
-class LocalVmRoute : public RouteData {
-public:
-    LocalVmRoute(const VmInterfaceKey &intf, uint32_t label,
-                 int tunnel_bmap, bool force_policy, const string &vn_name,
-                 uint8_t flags, const SecurityGroupList &sg_list,
-                 Op op = RouteData::CHANGE) :
-        RouteData(op, false), intf_(intf), 
-        label_(label), tunnel_bmap_(tunnel_bmap),
-        force_policy_(force_policy), dest_vn_name_(vn_name),
-        proxy_arp_(true), sync_route_(false), 
-        flags_(flags), sg_list_(sg_list) { };
-    virtual ~LocalVmRoute() { };
-    void DisableProxyArp() {proxy_arp_ = false;};
-    virtual string ToString() const {return "local VM";};;
-    virtual bool AddChangePath(AgentPath *path);
-    const SecurityGroupList &GetSecurityGroupList() const {return sg_list_;}; 
-
-private:
-    VmInterfaceKey intf_;
-    uint32_t label_;
-    int tunnel_bmap_;
-    bool force_policy_;
-    string dest_vn_name_;
-    bool proxy_arp_;
-    bool sync_route_;
-    uint8_t flags_;
-    SecurityGroupList sg_list_;
-    DISALLOW_COPY_AND_ASSIGN(LocalVmRoute);
-};
-
-class RemoteVmRoute : public RouteData {
-public:
-    RemoteVmRoute(const string &vrf_name, const Ip4Address &addr,
-                  uint32_t label, const string &dest_vn_name,
-                  int bmap, const SecurityGroupList &sg_list,
-                  DBRequest &req, Op op = RouteData::CHANGE):
-        RouteData(op, false), server_vrf_(vrf_name),
-        server_ip_(addr), tunnel_bmap_(bmap), 
-        label_(label), dest_vn_name_(dest_vn_name), sg_list_(sg_list)
-        {nh_req_.Swap(&req);}
-    virtual ~RemoteVmRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "remote VM";};;
-    const SecurityGroupList &GetSecurityGroupList() const {return sg_list_;}; 
-
-private:
-    string server_vrf_;
-    Ip4Address server_ip_;
-    int tunnel_bmap_;
-    uint32_t label_;
-    string dest_vn_name_;
-    SecurityGroupList sg_list_;
-    DBRequest nh_req_;
-    DISALLOW_COPY_AND_ASSIGN(RemoteVmRoute);
-};
-
-class InetInterfaceRoute : public RouteData {
-public:
-    InetInterfaceRoute(const InetInterfaceKey &intf, uint32_t label,
-                       int tunnel_bmap, const string &dest_vn_name,
-                       Op op = RouteData::CHANGE) : 
-        RouteData(op,false), intf_(intf), label_(label), 
-        tunnel_bmap_(tunnel_bmap), dest_vn_name_(dest_vn_name) { };
-    virtual ~InetInterfaceRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "host";};;
-
-private:
-    InetInterfaceKey intf_;
-    uint32_t label_;
-    int tunnel_bmap_;
-    string dest_vn_name_;
-    DISALLOW_COPY_AND_ASSIGN(InetInterfaceRoute);
-};
-
-class HostRoute : public RouteData {
-public:
-    HostRoute(const PacketInterfaceKey &intf, const string &dest_vn_name,
-              Op op  = RouteData::CHANGE) : 
-        RouteData(op, false), intf_(intf),
-        dest_vn_name_(dest_vn_name), proxy_arp_(false) { };
-    virtual ~HostRoute() { };
-    void EnableProxyArp() {proxy_arp_ = true;};
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "host";};;
-
-private:
-    PacketInterfaceKey intf_;
-    string dest_vn_name_;
-    bool proxy_arp_;
-    DISALLOW_COPY_AND_ASSIGN(HostRoute);
-};
-
-class VlanNhRoute : public RouteData {
-public:
-    VlanNhRoute(const VmInterfaceKey &intf, uint16_t tag, uint32_t label,
-                const string &dest_vn_name, const SecurityGroupList &sg_list,
-                Op op  = RouteData::CHANGE) :
-        RouteData(op, false), intf_(intf),
-        tag_(tag), label_(label), dest_vn_name_(dest_vn_name), 
-        sg_list_(sg_list) { };
-    virtual ~VlanNhRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "vlannh";};;
-
-private:
-    VmInterfaceKey intf_;
-    uint16_t tag_;
-    uint32_t label_;
-    string dest_vn_name_;
-    SecurityGroupList sg_list_;
-    DISALLOW_COPY_AND_ASSIGN(VlanNhRoute);
-};
-
-class MulticastRoute : public RouteData {
-public:
-    MulticastRoute(const Ip4Address &src_addr, 
-                   const Ip4Address &grp_addr,
-                   const string &vn_name, 
-                   const string &vrf_name,
-                   int vxlan_id,
-                   COMPOSITETYPE type, Op op  = RouteData::CHANGE) :
-        RouteData(op, true), 
-        src_addr_(src_addr), grp_addr_(grp_addr),
-        vn_name_(vn_name), vrf_name_(vrf_name), vxlan_id_(vxlan_id),
-        comp_type_(type) { };
-    virtual ~MulticastRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "multicast";};;
-
-private:
-    Ip4Address src_addr_;
-    Ip4Address grp_addr_;
-    string vn_name_;
-    string vrf_name_;
-    int vxlan_id_;
-    COMPOSITETYPE comp_type_;
-    DISALLOW_COPY_AND_ASSIGN(MulticastRoute);
-};
-
-class ReceiveRoute : public RouteData {
-public:
-    ReceiveRoute(const InetInterfaceKey &intf, uint32_t label,
-                 uint32_t tunnel_bmap, bool policy, const string &vn,
-                 Op op  = RouteData::CHANGE) : 
-        RouteData(op, false), intf_(intf), 
-        label_(label), tunnel_bmap_(tunnel_bmap),
-        policy_(policy), proxy_arp_(false), vn_(vn), sg_list_() {};
-    virtual ~ReceiveRoute() { };
-    void EnableProxyArp() {proxy_arp_ = true;};
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "receive";};;
-
-private:
-    InetInterfaceKey intf_;
-    uint32_t label_;
-    int tunnel_bmap_;
-    bool policy_;
-    bool proxy_arp_;
-    string vn_;
-    SecurityGroupList sg_list_;
-    DISALLOW_COPY_AND_ASSIGN(ReceiveRoute);
-};
-
-class Inet4UnicastEcmpRoute : public RouteData {
-public:
-    Inet4UnicastEcmpRoute(const Ip4Address &dest_addr, uint8_t plen,
-                          const string &vn_name, 
-                          uint32_t label, bool local_ecmp_nh, 
-                          const string &vrf_name, SecurityGroupList sg_list,
-                          DBRequest &nh_req, Op op  = RouteData::CHANGE) :
-        RouteData(op, false), dest_addr_(dest_addr), plen_(plen),
-        vn_name_(vn_name), label_(label), local_ecmp_nh_(local_ecmp_nh),
-        vrf_name_(vrf_name), sg_list_(sg_list) {
-            nh_req_.Swap(&nh_req);
-        };
-    virtual ~Inet4UnicastEcmpRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "inet4 ecmp";};;
-
-private:
-    Ip4Address dest_addr_;
-    uint8_t plen_;
-    string vn_name_;
-    uint32_t label_;
-    bool local_ecmp_nh_;
-    string vrf_name_;
-    SecurityGroupList sg_list_;
-    DBRequest nh_req_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4UnicastEcmpRoute);
-};
-
-
-class Inet4UnicastArpRoute : public RouteData {
-public:
-    Inet4UnicastArpRoute(const string &vrf_name, 
-                         const Ip4Address &addr,
-                         Op op  = RouteData::CHANGE) :
-        RouteData(op, false), vrf_name_(vrf_name), addr_(addr) { };
-    virtual ~Inet4UnicastArpRoute() { };
-
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "arp";};;
-private:
-    string vrf_name_;
-    Ip4Address addr_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4UnicastArpRoute);
-};
-
-class Inet4UnicastGatewayRoute : public RouteData {
-public:
-    Inet4UnicastGatewayRoute(const Ip4Address &gw_ip, 
-                             const string &vrf_name,
-                             Op op  = RouteData::CHANGE) :
-        RouteData(op, false), gw_ip_(gw_ip), 
-        vrf_name_(vrf_name) { };
-    virtual ~Inet4UnicastGatewayRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "gateway";};;
-
-private:
-    Ip4Address gw_ip_;
-    string vrf_name_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4UnicastGatewayRoute);
-};
-
-class DropRoute : public RouteData {
-public:
-    DropRoute(Op op  = RouteData::CHANGE) :
-        RouteData(op, false) { };
-    virtual ~DropRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "drop";};;
-private:
-    DISALLOW_COPY_AND_ASSIGN(DropRoute);
-};
-
-//////////////////////////////////////////////////////////////////
-//  UNICAST INET4
-/////////////////////////////////////////////////////////////////
-//TODO Remove this, flatten out
-class Inet4AgentRouteTable : public AgentRouteTable {
-public:
-    enum Type {
-        UNICAST,
-        MULTICAST,
-    };
-    Inet4AgentRouteTable(Type type, DB *db, const std::string &name) : 
-        AgentRouteTable(db, name), type_(type) { };
-    virtual ~Inet4AgentRouteTable() { };
-    virtual void ProcessDelete(RouteEntry *rt) { };
-    virtual void ProcessAdd(RouteEntry *rt) { };
-    virtual string GetTableName() const {return "Inet4AgentRouteTable";};
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::INET4_UNICAST;};
-    Type GetInetRouteType() { return type_; };
-
-private:
-    Type type_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4AgentRouteTable);
-};
-
-class Inet4UnicastRouteEntry : public RouteEntry {
-public:
-    Inet4UnicastRouteEntry(VrfEntry *vrf, const Ip4Address &addr) : 
-        RouteEntry(vrf, false), addr_(addr) { 
-            plen_ = 32; 
-        };
-    Inet4UnicastRouteEntry(VrfEntry *vrf, const Ip4Address &addr, 
-                           uint8_t plen, bool is_multicast) : 
-        RouteEntry(vrf, is_multicast), addr_(addr), plen_(plen) { 
-            addr_ = boost::asio::ip::address_v4(addr.to_ulong() & 
-                                    (plen ? (0xFFFFFFFF << (32 - plen)) : 0));
-        };
-    virtual ~Inet4UnicastRouteEntry() { };
-
-    virtual int CompareTo(const Route &rhs) const;
-    virtual string ToString() const;
-    virtual KeyPtr GetDBRequestKey() const;
-    virtual void SetKey(const DBRequestKey *key);
-    virtual void RouteResyncReq() const;
-    virtual bool DBEntrySandesh(Sandesh *sresp) const;
-    virtual const string GetAddressString() const {return addr_.to_string();};
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::INET4_UNICAST;};
-
-    bool DBEntrySandesh(Sandesh *sresp, 
-                        Ip4Address addr,
-                        uint8_t plen) const;
-
-    const Ip4Address &GetIpAddress() const {return addr_;};
-    int GetPlen() const {return plen_;};
-    void SetAddr(Ip4Address addr) { addr_ = addr; };
-    void SetPlen(int plen) { plen_ = plen; };
-    //Key for patricia node lookup 
-    class Rtkey {
-      public:
-          static std::size_t Length(RouteEntry *key) {
-              Inet4UnicastRouteEntry *uckey =
-                  static_cast<Inet4UnicastRouteEntry *>(key);
-              return uckey->GetPlen();
-          }
-          static char ByteValue(RouteEntry *key, std::size_t i) {
-              Inet4UnicastRouteEntry *uckey =
-                  static_cast<Inet4UnicastRouteEntry *>(key);
-              const Ip4Address::bytes_type &addr_bytes = 
-                  uckey->GetIpAddress().to_bytes();
-              return static_cast<char>(addr_bytes[i]);
-          }
-    };
-
-private:
-    friend class Inet4UnicastAgentRouteTable;
-
-    Ip4Address addr_;
-    uint8_t plen_;
-    Patricia::Node rtnode_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4UnicastRouteEntry);
-};
-
-class Inet4UnicastAgentRouteTable : public Inet4AgentRouteTable {
-public:
-    typedef Patricia::Tree<Inet4UnicastRouteEntry, &Inet4UnicastRouteEntry::rtnode_, 
-            Inet4UnicastRouteEntry::Rtkey> Inet4RouteTree;
-
-    Inet4UnicastAgentRouteTable(DB *db, const std::string &name) :
-        Inet4AgentRouteTable(Inet4AgentRouteTable::UNICAST, db, name), 
-        walkid_(DBTableWalker::kInvalidWalkerId) { };
-    virtual ~Inet4UnicastAgentRouteTable() { };
-
-    Inet4UnicastRouteEntry *FindLPM(const Ip4Address &ip);
-    virtual string GetTableName() const {return "Inet4UnicastAgentRouteTable";};
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::INET4_UNICAST;};
-    virtual void ProcessAdd(RouteEntry *rt) { 
-        tree_.Insert(static_cast<Inet4UnicastRouteEntry *>(rt));
-    };
-    virtual void ProcessDelete(RouteEntry *rt) { 
-        tree_.Remove(static_cast<Inet4UnicastRouteEntry *>(rt));
-    };
-    Inet4UnicastRouteEntry *FindRoute(const Ip4Address &ip) { 
-        return FindLPM(ip); };
-
-    Inet4UnicastRouteEntry *FindResolveRoute(const Ip4Address &ip);
-    static Inet4UnicastRouteEntry *FindResolveRoute(const string &vrf_name, 
-                                                    const Ip4Address &ip);
-    static DBTableBase *CreateTable(DB *db, const std::string &name);
-    static Inet4UnicastRouteEntry *FindRoute(const string &vrf_name, 
-                                             const Ip4Address &ip);
-    static void RouteResyncReq(const string &vrf_name, 
-                               const Ip4Address &ip, uint8_t plen);
-    static void DeleteReq(const Peer *peer, const string &vrf_name,
-                          const Ip4Address &addr, uint8_t plen);
-    static void Delete(const Peer *peer, const string &vrf_name,
-                       const Ip4Address &addr, uint8_t plen);
-    static void AddHostRoute(const string &vrf_name,
-                             const Ip4Address &addr, uint8_t plen,
-                             const std::string &dest_vn_name);
-    static void AddVlanNHRouteReq(const Peer *peer, const string &vm_vrf,
-                                  const Ip4Address &addr, uint8_t plen,
-                                  const uuid &intf_uuid, uint16_t tag,
-                                  uint32_t label, const string &dest_vn_name,
-                                  const SecurityGroupList &sg_list_);
-    static void AddVlanNHRoute(const Peer *peer, const string &vm_vrf,
-                               const Ip4Address &addr, uint8_t plen,
-                               const uuid &intf_uuid, uint16_t tag,
-                               uint32_t label, const string &dest_vn_name,
-                               const SecurityGroupList &sg_list_);
-   static void AddLocalVmRouteReq(const Peer *peer, const string &vm_vrf,
-                                  const Ip4Address &addr, uint8_t plen,
-                                  const uuid &intf_uuid,
-                                  const string &vn_name,
-                                  uint32_t label, bool force_policy);
-    static void AddLocalVmRoute(const Peer *peer, const string &vm_vrf,
-                                const Ip4Address &addr, uint8_t plen,
-                                const uuid &intf_uuid,
-                                const string &vn_name,
-                                uint32_t label, bool force_policy);
-    static void AddSubnetBroadcastRoute(const Peer *peer, 
-                                        const string &vrf_name,
-                                        const Ip4Address &src_addr, 
-                                        const Ip4Address &grp_addr,
-                                        const string &vn_name);
-    static void AddLocalVmRouteReq(const Peer *peer, const string &vm_vrf,
-                                   const Ip4Address &addr, uint8_t plen,
-                                   const uuid &intf_uuid, const string &vn_name,
-                                   uint32_t label, 
-                                   const SecurityGroupList &sg_list_);
-    static void AddLocalVmRoute(const Peer *peer, const string &vm_vrf,
-                                const Ip4Address &addr, uint8_t plen,
-                                const uuid &intf_uuid, const string &vn_name,
-                                uint32_t label, 
-                                const SecurityGroupList &sg_list_);
-    static void AddLocalVmRouteReq(const Peer *peer, const string &vm_vrf,
-                                   const Ip4Address &addr, uint8_t plen,
-                                   const uuid &intf_uuid, const string &vn_name,
-                                   uint32_t label);
-    static void AddRemoteVmRouteReq(const Peer *peer, const string &vm_vrf,
-                                    const Ip4Address &vm_addr,uint8_t plen,
-                                    const Ip4Address &server_ip,
-                                    TunnelType::TypeBmap bmap, uint32_t label,
-                                    const string &dest_vn_name);
-    static void AddRemoteVmRouteReq(const Peer *peer, const string &vm_vrf,
-                                    const Ip4Address &vm_addr, uint8_t plen,
-                                    const Ip4Address &server_ip,
-                                    TunnelType::TypeBmap bmap,uint32_t label,
-                                    const string &dest_vn_name,
-                                    const SecurityGroupList &sg_list_);
-    static void AddRemoteVmRouteReq(const Peer *peer, const string &vm_vrf,
-                                    const Ip4Address &vm_addr,uint8_t plen,
-                                    std::vector<ComponentNHData> comp_nh_list,
-                                    uint32_t label,
-                                    const string &dest_vn_name,
-                                    const SecurityGroupList &sg_list_);
-    static void AddLocalEcmpRoute(const Peer *peer, const string &vm_vrf,
-                                  const Ip4Address &vm_addr,uint8_t plen,
-                                  std::vector<ComponentNHData> comp_nh_list,
-                                  uint32_t label,
-                                  const string &dest_vn_name,
-                                  const SecurityGroupList &sg_list_);
-    static void CheckAndAddArpReq(const string &vrf_name, const Ip4Address &ip);
-    static void AddArpReq(const string &vrf_name, const Ip4Address &ip); 
-    static void ArpRoute(DBRequest::DBOperation op, 
-                         const Ip4Address &ip, 
-                         const struct ether_addr &mac,
-                         const string &vrf_name, 
-                         const Interface &intf,
-                         bool resolved,
-                         const uint8_t plen);
-    static void AddResolveRoute(const string &vrf_name, 
-                                const Ip4Address &ip, 
-                                const uint8_t plen); 
-    static void AddInetInterfaceRoute(const Peer *peer, const string &vm_vrf,
-                                      const Ip4Address &addr, uint8_t plen,
-                                      const string &interface, uint32_t label,
-                                      const string &vn_name);
-    static void AddVHostRecvRoute(const Peer *peer, const string &vrf,
-                                  const string &interface,
-                                  const Ip4Address &addr, uint8_t plen,
-                                  const string &vn_name, bool policy);
-    static void AddVHostRecvRouteReq(const Peer *peer, const string &vrf,
-                                     const string &interface,
-                                     const Ip4Address &addr, uint8_t plen,
-                                     const string &vn_name, bool policy);
-    static void AddVHostSubnetRecvRoute(const Peer *peer, const string &vrf,
-                                        const string &interface,
-                                        const Ip4Address &addr, uint8_t plen,
-                                        const std::string &vn_name,
-                                        bool policy);
-    static void DelVHostSubnetRecvRoute(const string &vm_vrf,
-                                        const Ip4Address &addr, uint8_t plen);
-    static void AddDropRoute(const string &vm_vrf,
-                             const Ip4Address &addr, uint8_t plen);
-    static void AddGatewayRoute(const string &vrf_name,
-                                const Ip4Address &dst_addr,uint8_t plen,
-                                const Ip4Address &gw_ip,
-                                const std::string &vn_name);
-    static void AddGatewayRouteReq(const string &vrf_name,
-                                   const Ip4Address &dst_addr,uint8_t plen,
-                                   const Ip4Address &gw_ip,
-                                   const std::string &vn_name);
-
-private:
-    Inet4RouteTree tree_;
-    Patricia::Node rtnode_;
-    DBTableWalker::WalkId walkid_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4UnicastAgentRouteTable);
-};
-
-class Inet4UnicastRouteKey : public RouteKey {
-public:
-    Inet4UnicastRouteKey(const Peer *peer, const string &vrf_name, 
-                         const Ip4Address &dip, uint8_t plen) : 
-        RouteKey(peer, vrf_name), dip_(dip), plen_(plen) { };
-    virtual ~Inet4UnicastRouteKey() { };
-    //Called from oute creation in input of route table
-    virtual RouteEntry *AllocRouteEntry(VrfEntry *vrf, bool is_multicast) const;
-    //Enqueue add/chg/delete for route
-    virtual AgentRouteTable *GetRouteTableFromVrf(VrfEntry *vrf) { 
-        return (static_cast<AgentRouteTable *>(vrf->
-                         GetRouteTable(AgentRouteTableAPIS::INET4_UNICAST)));
-    };
-    virtual AgentRouteTableAPIS::TableType GetRouteTableType() {
-       return AgentRouteTableAPIS::INET4_UNICAST;
-    }; 
-    virtual string ToString() const { return ("Inet4UnicastRouteKey"); };
-
-    const Ip4Address &GetAddress() const {return dip_;};
-    const uint8_t &GetPlen() const {return plen_;};
-
-private:
-    Ip4Address dip_;
-    uint8_t plen_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4UnicastRouteKey);
-};
-
-//////////////////////////////////////////////////////////////////
-//  MULTICAST INET4
-/////////////////////////////////////////////////////////////////
-class Inet4MulticastAgentRouteTable : public Inet4AgentRouteTable {
-public:
-    Inet4MulticastAgentRouteTable(DB *db, const std::string &name) :
-        Inet4AgentRouteTable(Inet4AgentRouteTable::MULTICAST, db, name),
-        walkid_(DBTableWalker::kInvalidWalkerId) { };
-    virtual ~Inet4MulticastAgentRouteTable() { };
-    virtual bool DelExplicitRoute(DBTablePartBase *part, DBEntryBase *entry) { 
-        return true; };
-    //Nexthop will be stored in path as lcoalvmpeer peer so that it falls in line
-    //Override virtual routines for no action w.r.t. multicast
-    virtual string GetTableName() const {return "Inet4MulticastAgentRouteTable";};
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::INET4_MULTICAST;};
-
-    void McRtRouteNotifyDone(DBTableBase *base, DBState *);
-    void AddVHostRecvRoute(const string &vm_vrf,
-                           const string &interface_name,
-                           const Ip4Address &addr,
-                           bool policy);
-    static DBTableBase *CreateTable(DB *db, const std::string &name);
-    static Inet4MulticastRouteEntry *FindRoute(const string &vrf_name, 
-                                               const Ip4Address &ip, 
-                                               const Ip4Address &dip);
-    static void RouteResyncReq(const string &vrf_name, 
-                               const Ip4Address &sip, 
-                               const Ip4Address &dip);
-    static void AddMulticastRoute(const string &vrf_name, 
-                                  const string &vn_name,
-                                  const Ip4Address &src_addr,
-                                  const Ip4Address &grp_addr);
-    static void DeleteMulticastRoute(const string &vrf_name, 
-                                     const Ip4Address &src_addr,
-                                     const Ip4Address &grp_addr); 
-    static void Delete(const string &vrf_name, const Ip4Address &src_addr,
-                       const Ip4Address &grp_addr);
-private:
-    DBTableWalker::WalkId walkid_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4MulticastAgentRouteTable);
-};
-
-class Inet4MulticastRouteEntry : public RouteEntry {
-public:
-    Inet4MulticastRouteEntry(VrfEntry *vrf, const Ip4Address &dst, 
-                             const Ip4Address &src) :
-        RouteEntry(vrf, true), dst_addr_(dst), src_addr_(src) { }; 
-    virtual ~Inet4MulticastRouteEntry() { };
-
-    virtual int CompareTo(const Route &rhs) const;
-    virtual string ToString() const;
-    virtual KeyPtr GetDBRequestKey() const;
-    virtual void SetKey(const DBRequestKey *key);
-    virtual void RouteResyncReq() const;
-    virtual const string GetAddressString() const {
-        return dst_addr_.to_string();};
-    virtual bool DBEntrySandesh(Sandesh *sresp) const;
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::INET4_MULTICAST;};;
-
-    void SetDstIpAddress(const Ip4Address &dst) {dst_addr_ = dst;};
-    void SetSrcIpAddress(const Ip4Address &src) {src_addr_ = src;};
-    const Ip4Address &GetSrcIpAddress() const {return src_addr_;};
-    const Ip4Address &GetDstIpAddress() const {return dst_addr_;};
-
-private:
-    Ip4Address dst_addr_;
-    Ip4Address src_addr_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4MulticastRouteEntry);
-};
-
-class Inet4MulticastRouteKey : public RouteKey {
-public:
-    Inet4MulticastRouteKey(const string &vrf_name,const Ip4Address &dip, 
-                           const Ip4Address &sip) :
-                         RouteKey(Agent::GetInstance()->GetLocalVmPeer(), 
-                                  vrf_name), dip_(dip), sip_(sip) { };
-    Inet4MulticastRouteKey(const string &vrf_name, const Ip4Address &dip) : 
-        RouteKey(Agent::GetInstance()->GetLocalVmPeer(), vrf_name), dip_(dip) { 
-            boost::system::error_code ec;
-            sip_ =  IpAddress::from_string("0.0.0.0", ec).to_v4();
-    };
-    Inet4MulticastRouteKey(const string &vrf_name) : 
-        RouteKey(Agent::GetInstance()->GetLocalVmPeer(), vrf_name) { 
-            boost::system::error_code ec;
-            dip_ =  IpAddress::from_string("255.255.255.255", ec).to_v4();
-            sip_ =  IpAddress::from_string("0.0.0.0", ec).to_v4();
-    };
-    virtual ~Inet4MulticastRouteKey() { };
-    virtual RouteEntry *AllocRouteEntry(VrfEntry *vrf, bool is_multicast) const;
-    //Enqueue add/chg/delete for route
-    virtual AgentRouteTable *GetRouteTableFromVrf(VrfEntry *vrf) { 
-        return (static_cast<Inet4MulticastAgentRouteTable *>
-                (vrf->GetRouteTable(AgentRouteTableAPIS::INET4_MULTICAST)));
-    };
-    virtual AgentRouteTableAPIS::TableType GetRouteTableType() {
-       return AgentRouteTableAPIS::INET4_MULTICAST;
-    }; 
-    virtual string ToString() const { return ("Inet4MulticastRouteKey"); };
-
-    const Ip4Address &GetDstIpAddress() const {return dip_;};
-    const Ip4Address &GetSrcIpAddress() const {return sip_;};
-
-private:
-    Ip4Address dip_;
-    Ip4Address sip_;
-    DISALLOW_COPY_AND_ASSIGN(Inet4MulticastRouteKey);
-};
-
-//////////////////////////////////////////////////////////////////
-//  LAYER2
-/////////////////////////////////////////////////////////////////
-class Layer2AgentRouteTable : public AgentRouteTable {
-public:
-    Layer2AgentRouteTable(DB *db, const std::string &name) : 
-        AgentRouteTable(db, name) { };
-    virtual ~Layer2AgentRouteTable() { };
-
-    virtual string GetTableName() const {return "Layer2AgentRouteTable";};
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::LAYER2;};
-    virtual void RouteTableWalkerNotify(VrfEntry *vrf, AgentXmppChannel *xmpp, 
-                                        DBState *state, bool associate,
-                                        bool unicast_walk, bool multicast_walk) {
-        //Dont support multicast walk
-        AgentRouteTable::RouteTableWalkerNotify(vrf, xmpp, state, associate, 
-                                                unicast_walk, false);
-    };
-
-    static void RouteResyncReq(const string &vrf_name, 
-                               const struct ether_addr &mac);
-    static DBTableBase *CreateTable(DB *db, const std::string &name);
-    static Layer2RouteEntry *FindRoute(const string &vrf_name, 
-                                       const struct ether_addr &mac);
-    static void AddRemoteVmRouteReq(const Peer *peer,
-                                    const string &vrf_name,
-                                    TunnelType::TypeBmap bmap,
-                                    const Ip4Address &server_ip,
-                                    uint32_t label,
-                                    struct ether_addr &mac,
-                                    const Ip4Address &vm_ip, uint32_t plen);
-    static void AddLocalVmRouteReq(const Peer *peer,
-                                   const uuid &intf_uuid,
-                                   const string &vn_name, 
-                                   const string &vrf_name,
-                                   uint32_t label, int bmap, 
-                                   struct ether_addr &mac,
-                                   const Ip4Address &vm_ip,
-                                   uint32_t plen); 
-    static void AddLocalVmRoute(const Peer *peer,
-                                const uuid &intf_uuid,
-                                const string &vn_name, 
-                                const string &vrf_name,
-                                uint32_t label, int bmap, 
-                                struct ether_addr &mac,
-                                const Ip4Address &vm_ip,
-                                uint32_t plen); 
-    static void AddLayer2BroadcastRoute(const string &vrf_name,
-                                        const string &vn_name,
-                                        const Ip4Address &dip,
-                                        const Ip4Address &sip,
-                                        int vxlan_id);
-    static void DeleteReq(const Peer *peer, const string &vrf_name,
-                          const struct ether_addr &mac);
-    static void Delete(const Peer *peer, const string &vrf_name,
-                       const struct ether_addr &mac);
-    static void DeleteBroadcastReq(const string &vrf_name);
-private:
-    DBTableWalker::WalkId walkid_;
-    DISALLOW_COPY_AND_ASSIGN(Layer2AgentRouteTable);
-};
-
-class Layer2RouteEntry : public RouteEntry {
-public:
-    Layer2RouteEntry(VrfEntry *vrf, const struct ether_addr &mac,
-                     const Ip4Address &vm_ip, uint32_t plen,
-                     Peer::Type type, bool is_multicast) : 
-        RouteEntry(vrf, is_multicast), mac_(mac){ 
-        if (type != Peer::BGP_PEER) {
-            vm_ip_ = vm_ip;
-            plen_ = plen;
-        } else {
-            //TODO Add the IP prefix sent by BGP peer to add IP route 
-        }
-    };
-    virtual ~Layer2RouteEntry() { };
-
-    virtual int CompareTo(const Route &rhs) const;
-    virtual string ToString() const;
-    virtual void UpdateDependantRoutes() { };
-    virtual void UpdateNH() { };
-    virtual KeyPtr GetDBRequestKey() const;
-    virtual void SetKey(const DBRequestKey *key);
-    virtual void RouteResyncReq() const;
-    virtual const string GetAddressString() const {
-        //For multicast use the same tree as of 255.255.255.255
-        if (IsMulticast()) 
-            return "255.255.255.255";
-        return ToString();
-    };
-    virtual AgentRouteTableAPIS::TableType GetTableType() const {
-        return AgentRouteTableAPIS::LAYER2;};;
-    virtual bool DBEntrySandesh(Sandesh *sresp) const;
-
-    const struct ether_addr &GetAddress() const {return mac_;};
-    const Ip4Address &GetVmIpAddress() const {return vm_ip_;};
-    const uint32_t GetVmIpPlen() const {return plen_;};
-
-private:
-    struct ether_addr mac_;
-    Ip4Address vm_ip_;
-    uint32_t plen_;
-    DISALLOW_COPY_AND_ASSIGN(Layer2RouteEntry);
-};
-
-class Layer2RouteKey : public RouteKey {
-public:
-    Layer2RouteKey(const Peer *peer, const string &vrf_name, 
-                   const struct ether_addr &mac) :
-        RouteKey(peer, vrf_name), dmac_(mac) { };
-    Layer2RouteKey(const Peer *peer, const string &vrf_name, 
-                   const struct ether_addr &mac, const Ip4Address &vm_ip,
-                   uint32_t plen) :
-        RouteKey(peer, vrf_name), dmac_(mac), vm_ip_(vm_ip), plen_(plen) { };
-    Layer2RouteKey(const Peer *peer, const string &vrf_name) : 
-        RouteKey(peer, vrf_name) { 
-            dmac_ = *ether_aton("FF:FF:FF:FF:FF:FF");
-        };
-    virtual ~Layer2RouteKey() { };
-
-    virtual RouteEntry *AllocRouteEntry(VrfEntry *vrf, bool is_multicast) const;
-    //Enqueue add/chg/delete for route
-    virtual AgentRouteTable *GetRouteTableFromVrf(VrfEntry *vrf) { 
-        return (static_cast<AgentRouteTable *>(vrf->
-                            GetRouteTable(AgentRouteTableAPIS::LAYER2)));
-    };
-    virtual AgentRouteTableAPIS::TableType GetRouteTableType() {
-        return AgentRouteTableAPIS::LAYER2;
-    }; 
-    virtual string ToString() const { return ("Layer2RouteKey"); };
-    const struct ether_addr &GetMac() const { return dmac_;};
-
-private:
-    struct ether_addr dmac_;
-    Ip4Address vm_ip_;
-    uint32_t plen_;
-    DISALLOW_COPY_AND_ASSIGN(Layer2RouteKey);
-};
-
-class Layer2EcmpRoute : public RouteData {
-public:
-    Layer2EcmpRoute(const struct ether_addr &dest_addr, 
-                    const string &vn_name, 
-                    const string &vrf_name, uint32_t label, 
-                    bool local_ecmp_nh,
-                    Op op  = RouteData::CHANGE) : 
-        RouteData(op , false), dest_addr_(dest_addr),
-        vn_name_(vn_name), vrf_name_(vrf_name),
-        label_(label), local_ecmp_nh_(local_ecmp_nh) {
-        };
-    virtual ~Layer2EcmpRoute() { };
-    virtual bool AddChangePath(AgentPath *path);
-    virtual string ToString() const {return "layer2ecmp";};;
-
-private:
-    const struct ether_addr dest_addr_;
-    string vn_name_;
-    string vrf_name_;
-    uint32_t label_;
-    bool local_ecmp_nh_;
-    DISALLOW_COPY_AND_ASSIGN(Layer2EcmpRoute);
 };
 
 #define AGENT_DBWALK_TRACE_BUF "AgentDBwalkTrace"
