+++ src/vnsw/agent/oper/global_vrouter.cc
@@ -13,7 +13,7 @@
 #include <oper/nexthop.h>
 #include <oper/vn.h>
 #include <oper/vrf.h>
-#include <oper/agent_route.h>
+#include <oper/route_common.h>
 #include <oper/operdb_init.h>
 #include <oper/global_vrouter.h>
 #include <base/util.h>
