+++ src/vnsw/agent/oper/interface.cc
@@ -14,7 +14,7 @@
 #include <cfg/cfg_init.h>
 #include <cfg/cfg_interface.h>
 #include <oper/operdb_init.h>
-#include <oper/agent_route.h>
+#include <oper/route_common.h>
 #include <oper/vm.h>
 #include <oper/vn.h>
 #include <oper/vrf.h>
