+++ src/vnsw/agent/oper/route_common.h
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2013 Juniper Networks, Inc. All rights reserved.
+ */
+
+#ifndef vnsw_route_common_hpp
+#define vnsw_route_common_hpp
+
+#include <oper/agent_route.h>
+#include <oper/agent_path.h>
+#include <oper/inet4_unicast_route.h>
+#include <oper/inet4_multicast_route.h>
+#include <oper/layer2_route.h>
+
+#endif // vnsw_route_common_hpp
