+++ src/vnsw/agent/pkt/flow_table.h
@@ -26,7 +26,7 @@
 #include <oper/vm.h>
 #include <oper/interface_common.h>
 #include <oper/nexthop.h>
-#include <oper/agent_route.h>
+#include <oper/route_common.h>
 
 class FlowStatsCollector;
 class PktSandeshFlow;
