+++ src/vnsw/agent/pkt/pkt_handler.cc
@@ -12,7 +12,7 @@
 #include "cmn/agent_stats.h"
 #include "oper/interface_common.h"
 #include "oper/nexthop.h"
-#include "oper/agent_route.h"
+#include "oper/route_common.h"
 #include "oper/vrf.h"
 #include "pkt/tap_interface.h"
 #include "pkt/test_tap_interface.h"
